Add support for /away on ScreenSaver activation.
Use NSDistributedNotificationCenter and the screensaver start/stop
events to set /away when the screensaver starts. The screensaver
events are undocumented, but the coupling is loose and only the
away-on-screensaver functionality is affected if Apple changes or
removes these out from under us.

The code is from the Adium project's AdiumIdleManager by
Evan Schoenberg.