Avoid some "view clipping content" warnings (rejig GUI).
The .xib files for non-en locales generate a bunch of warnings
from CompileXIB because the translated strings are too big to fit
in the alloted space. This resizes as many as trivially possible
to reduce the number of such warnings. To get rid of all of them
would require both major UI overhaul *and* optimizing the
translated strings, but this picks the low-hanging fruit.