Rewrite tab-completion to use selection rather than marked text.
Mac OS X 10.5 and 10.6 has a bug that eats insertTab: events on
NSTextFields when there is marked text in the field. Rewrite the
relevant bits of code to use selected text instead of marked text.
Took the opportunity to clean up the code in general.

This fixes Issue #3 on the Github tracker.