Set Base SDK build setting to "macosx"
The Base SDK should govern which version of the SDK is used, but all versions
of the SDK should support deploying to multiple OS versions (depending on the
value of the Deployment Target), so we'll set the Base SDK to "macosx" (i.e.
the latest SDK for the version of Xcode you're running) to get the latest
and greatest bug(fixe)s.