Drop last use of Carbon (for getting modifier key info on launch) in fe-aqua
During launch we check for modifier keys, and if the Option (Alt) key is down
we don't automatically connect to the network. This used to require using the
Carbon GetCurrentKeyModifiers() function, but as of OS X 10.6 NSEvent
provides a +modifierFlags (class) method that gives you the same information
using AppKit.