Reimplement SGFileUtility as XAFileUtil and use modern APIs
The code in SGFileUtility was using older file APIs or POSIX APIs and
contained no source code comments. XAFileUtil replaces this code with modern
APIs, copiously commented, and cleans up client code to be shorter, more
consistent, and with (somewhat) better error handling (still more to do on
that, but this is a start).

In particular, the native data type for XAFileUtil is now NSURLs rather than
the NSString/path-based SGFileUtility. Apple is moving in this direction too,
but they haven't completed the transition: there are still path-based APIs
with no URL-based counterparts, and several of the URL-based APIs are not
available in older versions of OS X (i.e. 10.6).