Minimal implementation of full screen support
This enables Full Screen support on the ChatWindow.xib, and adds an
"Enter Full Screen" item to the View menu. The former controls whether the
window widget in the top right corner for entering full screen is shown, and
the latter lets you trigger the toggle from a menu item (that you can assign
a keyboard shortcut to).

The full screen toggle is ignored when the runtime doesn't support full screen
(i.e. on Mac OS X 10.6), and the menu item should be automatically disabled on
older versions of Mac OS X since they do not respondToSelector: for the
relevant method. The menu item will still be visible, but should be inactive.