Implements sending events to the notification center.
As with growl the user can select in the options whether to send events to the
notification center and control whether that happens if the app is in the
foreground or always.

With this patch if the user has chosen to only display
notifications when in the background and the app is in the foreground the event
will not show a popup bubble but will still appear in the notification center
history.