Avoid using ancient OS X OpenSSL
This has been deprecated for ages, and it's removed in the 10.12 SDK.
Instead, statically link against OpenSSL from Homebrew, and bundle its certs.

Pods' OpenSSL is useless to us, since it's only built for iOS.

Optimally, we'd switch to Secure Transport, but that's a lot of work I'm not
interested in doing just yet.